+++ Projects/Services/NetworkModule/Sources/Base/BaseRemoteDataSource.swift
@@ -75,6 +75,7 @@ private extension BaseRemoteDataSource {
 
     func checkTokenIsExpired() -> Bool {
         let expired = keychain.load(type: .accessExpiredAt).toDMSDate()
+        print(Date(), expired)
         return Date() > expired
     }
