+++ Projects/Services/WatchRestAPIModule/Sources/Base/Plugin/JwtPlugin.swift
@@ -21,20 +21,6 @@ public struct JwtPlugin: PluginType {
         req.addValue(token, forHTTPHeaderField: jwtTokenType.rawValue)
         return req
     }
-
-    public func didReceive(
-        _ result: Result<Response, MoyaError>,
-        target: TargetType
-    ) {
-        switch result {
-        case let .success(res):
-            if let new = try? res.map(TokenDTO.self) {
-                saveToken(token: new)
-            }
-        default:
-            break
-        }
-    }
 }
 
 private extension JwtPlugin {
